requirements.txt
langchain==0.1.0
langchain-community==0.0.20
chromadb==0.4.22
pypdf==3.17.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
llama-cpp-python==0.2.27
transformers==4.30.2
huggingface-hub==0.16.4
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.10.1 